Redone Website started from scratch
I started out with making new js, css, html and adding grid.css and
type.css. I designed all these in adobe illustrator and saved them as
svg these pieces were all brained stormed through sketching on paper
and I used it as my websites example. This website will be reflecting
my photography skill and portfolio pieces

The old website since it was a graphic it was hard to implement a
navigation bar on top and the website generally had nothing to do with
my college work, portfolio it was simply because of hobby of liking
sport cars.