Added The rest of the javascript
Added three buttons which would open and explain what Perception,
shutter speed and flash when button is pressed it pops out a box which
is a border within it provides a description of the topics. Used same
javascript code we used in the the assignment 6 the hide/show box and
created 3 buttons and box each with different names. Added the
javascript.