Stylized boxes
Added background colour for boxes and added height and width added a
font from google fonts and resized font. Grid would be useful but It
sort of glitches because my 4 transitions interferes because it's using
its own grid system so I have to try to balance all the code and use
web query.