Repositioned
Added grid on the buttons and text boxes and centred them to match the
icons on the top I stylized them added border radius and margin and
colour orange. Made the billboard calc 70vh instead of 100vh because it
made the billboard push the buttons far down. I also changed the nav
button transition colour when you hover over it changes to navajowhite
which is a good colour where people can see the colour change where as
black with a black nav bar wasn’t great.