Created smooth scrolling
I had a lot of issues going when letting the icons be pressed and then
they take you to the description sections in the bottom I basically
used a certain section connect to the bottom footer. I created the
scroll buttons in adobe illustrator. This is just a one scroll to the
footer and then back to nav.