Added flash
I added opacity and transition to css when you scroll down the camera
flashes and this happens with javascript. The problem was the it was
not linking to page it just which is the scroll. To solve the issue I
went to Thomas he went to {offset: '60%', context: and added page which
is the class of my scrolling element.